requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
openai>=1.12.0
mem0ai>=0.0.20
chromadb>=0.4.22
python-dotenv>=1.0.0
httpx>=0.27.0
